Add Symbols module and changed symbols to match
The previous symbols were all changed to the same circle symbol, just
colored differently to prevent the board from changing due to the use of
different unicode circles.

The @board in #Board class was also changed to create 6 arrays, each
with 7 empty circle elements to make it easier to print to the console.