Add methods and tests to place a symbol on board
The #place_symbol, #place_column_number, and #verify_player_number
public methods were added to the Game class. Tests were also added for
all of these classes.

The Board class also has a #place method with no functionality at the
moment, but is meant to update the board and put the correct symbol in
the correct place on the board.