Add tests and methods for changing the board
Added #times_a_column_was_picked and #player_symbol methods to Game with
their respective tests.

Change #place_symbol to use some of the new methods to make passing
information to the Board object easier.

The Board object's initialization was changed to remove the passing of
Player objects to Board. Instead, if the Board object needs access to
player information, the Game object will pass that information directly.

The Board #place method was renamed to #update_board to help prevent any
confusion with some of the Game methods that also use place in their
name. Tests and functionality were also added for the method.