Add another condition to #place_column_number
We can use #times_a_column_was_picked to make sure that the same column
isn't picked more than 6 times since that's the number of rows this game
of Connect Four has, therefore that's how many symbols would be needed
to fill out that column.

A test was added to #place_column_number to test that a number cannot be
picked more than 6 times.