Add #check_for_game_over and other methods to play
The #check_for_game_over method handles both the ending and continuation
of the game, so relevant parts of #place_symbol were moved here instead.
Now, #place_symbol makes a call to #check_for_game_over instead
of #play_round to determine whether the game ends or continues,
and #play_round was simplified.

A #restart method was also added, but all it really does is create a new
Game instance in order to restart the game.

The Board's #update_board method was also changed slightly so that the
argument names are in the same order as they are used to update the
board. So, the arguments are now listed as row before column instead of
the other way around.

Board also had 2 new methods added to help look for any lines of 4 or
more to end the game.

The #check_rows method calls #check_transformation to handle looking for
a line of 4 or more of the same symbol in all 8 directions starting from
a given point on the board.

The #check_transformation method checks only 1 out of 8 possible
directions a line can be formed from a single point on the board.

All relevant tests were either added or updated to reflect the changes
made.