Fix board's #check_rows and refactor player inputs
The #check_rows method had an unnecessary condition that was causing the
program to break, so it was removed.

The 3 player input methods were consolidated into 1 method that only
gets the player's input and all the messages were moved to other
methods. All references to the old player input methods were changed for
all tests.